More cleanup, reorganizing
