+++ app/javascript/components/editor-config/images.js
