+++ app/javascript/components/editor-config/menu.js
@@ -5,7 +5,7 @@ import { redo, undo } from 'prosemirror-history'
 import { wrapInList } from './schema-list'
 // import { addColumnAfter, addColumnBefore } from 'prosemirror-tables'
 import { addAnnotation } from './plugin-comment'
-import { addFigure } from './image-plugin'
+import { addFigure } from './images'
 import schema from './schema'
 import icons from './icons'
